+++ src/Ui/Bundle/EventBundle/Controller/Happening/Broadcast/OpenStreamToPublicAction.php
@@ -4,27 +4,28 @@ namespace Proximum\Vimeet\Ui\Bundle\EventBundle\Controller\Happening\Broadcast;
 
 use Proximum\Vimeet\Application\Adapter\AuthorizationCheckerAdapterInterface;
 use Proximum\Vimeet\Application\Adapter\CommandBusInterface;
-use Proximum\Vimeet\Application\Command\Happening\Webinar\Broadcast\StartBroadcast;
-use Proximum\Vimeet\Application\Command\Happening\Webinar\Broadcast\StopBroadcast;
+use Proximum\Vimeet\Application\Command\Happening\Webinar\Broadcast\OpenStreamToPublicCommand;
 use Proximum\Vimeet\Application\Query\Happening\Webinar\CanAccessToWebinar;
 use Proximum\Vimeet\Domain\Model\Happening;
+use Proximum\Vimeet\Infrastructure\Adapter\CommandBus;
 use Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Security\Voter\Happening\BroadcastVoter;
 use Proximum\Vimeet\Ui\Bundle\EventBundle\ParamConverter\EventDomain;
 use Proximum\Vimeet\Ui\Bundle\EventBundle\ValueResolver\UserDomain;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
-class StreamAction
+class OpenStreamToPublicAction
 {
-    /** @var CommandBusInterface */
+    /** @var CommandBus */
     private $commandBus;
-
-    /** @var AuthorizationCheckerAdapterInterface */
+    /**
+     * @var AuthorizationCheckerAdapterInterface
+     */
     private $authorizationCheckerAdapter;
-
-    /** @var CanAccessToWebinar */
+    /**
+     * @var CanAccessToWebinar
+     */
     private $canAccessToWebinar;
 
     public function __construct(
@@ -50,32 +51,15 @@ class StreamAction
             || !$this->authorizationCheckerAdapter->isGranted('PERMISSION_HAPPENING_ACCESS', $event)
             || !$this->canAccessToWebinar->isSatisfiableBy($happening, $user)
             || $happening->getEvent() !== $event
-            || !$this->authorizationCheckerAdapter->isGranted(BroadcastVoter::CAN_STREAM_BROADCAST, $happening)
         ) {
             throw new AccessDeniedException('Access denied.');
         }
 
-        $message = null;
-        $type = $request->request->get('type');
-        $streamId = $request->request->get('streamId');
+        $mediaSharingType = $request->request->get('mediaSharingType');
+        $mediaSharingStream = $request->request->get('mediaSharingStream');
 
-        switch ($request->request->get('action')) {
-            case 'start':
-                $this->commandBus->handle(new StartBroadcast($happening, $type, $streamId));
-                $message = 'stream_started';
-            break;
-            case 'stop':
-                $message = 'stream_stopped';
-                $this->commandBus->handle(new StopBroadcast($happening, $type, $streamId));
-            break;
-            default:
-                throw new BadRequestHttpException('Unsupported action');
-            break;
-        }
+        $this->commandBus->handle(new OpenStreamToPublicCommand($happening, $mediaSharingType, $mediaSharingStream));
 
-        return new JsonResponse([
-            'status' => 'ok',
-            'message' => $message,
-        ]);
+        return new JsonResponse(['status' => 'ok',]);
     }
 }